// src/api/generated/models/DocAttributeDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* eslint-disable */

/**
 * Атрибут документа
 */
export type DocAttributeDto = {
  /**
   * ID атрибута
   */
  id?: number;
  /**
   * Наименование атрибута
   */
  name?: string;
  /**
   * Тип атрибута
   */
  type?: string;
};