// src/api/generated/models/UserRatingDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* eslint-disable */

/**
 * Рейтинг пользователя
 */
export type UserRatingDto = {
  /**
   * ID пользователя
   */
  id?: number;
  /**
   * Фамилия
   */
  lastName?: string;
  /**
   * Имя
   */
  firstName?: string;
  /**
   * Email
   */
  email?: string;
  /**
   * Количество созданных документов
   */
  createdDocuments?: number;
};